fix: Don't throw exception when trainee not found
A trainee not being found in the trainee details service is a valid
use case, in such scenarios we simply do not process the record.
Currently when a 404 is returned an exception is throw, which breaks the
kinesis stream updates. Catch 404s and simply log the error so that sync
can continue.

TISNEW-5344