fix: Do not send duplicate request for data
Requests are sent for every object not found, regardless of whether it
has already been requested. Keep track of the IDs that have been
requested and skip the request if it is already pending.
When a record is received for syncing remove it from the requested ID
list if found.

TIS21-1032