feat: throw any Exceptions from trainee details
As we filter based on role locally, there is no need to do this by catching HTTP 404s

refactor: use the trainee id from the converted DTO

The trainee ID might be in the "id" field or it might be somewhere else
i.e. For Qualification, it is actually in the "personId" field.

TIS21-1032: Synchronise placement details