chore: Provide region during workflow tests
The region provider chain must be able to find a region to initialize
several of the AWS beans without mocking them all.
The region should be able to be configured for tests using the
application property `cloud.aws.region.static` but the property is not
working as it is meant to.
As a workaround set the `AWS_REGION` environment variable when running
tests in the pipeline.

TIS21-1665