fix: Handle non-data DMS records
Some DMS records are sent with a record type of `control`, these records
notify of important events in the stream such as creating a table.

Control records contain no `data` so the record data map must be allowed
to be null.

Nothing needs to be done when receiving a control record, so processing
should be skipped but it should not cause an exception.

TIS21-1919