feat: Expand Specialty to PlacementSpecialty queue
When a specialty is updated the enricher gets all related placement
specialties and attempts to process them together, due to the amount of
time it can take to process large specialties this approach can cause
issues.
Placement specialties related to the updated specialty should instead be
queued up in the placement specialty queue, which in turn will handle
the associated placements.

TIS21-1692