feat: purge data associated with deleted records
When a `DELETE` message is received for Specialty or Trust, the
associated PlacementSpecialtys or Posts should also be deleted from the
sync lookup tables.

Remove `onAfterDelete` methods from Post and Site event listeners, they
are directly associated with placements and deleting the placement would
force deletion from the trainee profile. Removing a placement from the
trainee profile should only be done when the placement itself is
deleted.

TIS21-2549