feat: remove enrichment on placementspecialty delete
When a placement specialty is deleted additional checks are made on
whether the associated placement is already deleted, or whether the
placement already has a new specialty link.
If these are not true then the placement is re-enriched, requesting any
missing data as part of the process.

While there is no guarantee of order received, there is an expectation
that the sync service would eventually get any missing messages, either
the placement deletion or the new placement specialty.

The existing re-enrichment functionality on delete should be removed for
consistency with other related data, which are currently just deleted
without any affect on the placement - deletion nor enrichment.
Any handling of similar scenarios should consider all data deletions
that may affect a placement and ensure consistency for all.

TIS21-2549