fix: no Placement request for PlacementSpecialty
When handling PlacementSpecialty events the listener requests the
associated Placement if it is not found in the lookup data.
When doing a full load on an empty database this causes a large number
of requests, as no matter which order Placement and PlacementSpecialty
are loaded the other can not yet exist and will always be requested.

Removing the request from the PlacementSpecialty event listener allows
the PlacementSpecialty lookup data to be loaded prior to the Placements.

In the standard use case of a PlacementSpecialty being created, it is
fair to expect that the Placement will already exist in the lookup data,
or else the Placement insert message would be recieved eventually.

The Placement should still have the ability to request
PlacementSpecialty records should none exist in the lookup.

TIS21-2549