feat: log ECS metadata with X-Ray and Sentry
The ECS metadata endpoint provides information about what and where an
instance of this service is running. Including that information on
Sentry error captures allows easy identification of where an issue is
occuring.

Update the Sentry configurations to include the ECS metadata.
When not running in an ECS environment (i.e. the metadata endpoint is
not provided) the Sentry reports should not include extra metadata
e.g. showing null values.

TIS21-4443
TIS21-4379