refactor: ProgrammeMembership entity
Refactor the ProgrammeMembership entity to be a proper concrete POJO
instead of an empty extension of the `Record` POJO.
Create a mapper to convert between the Record's data map the the new
POJO.

Refactor the repository, service, event listeners and enricher for
ProgrammeMembership to use the new POJO. The enricher in particular
needs a lot of work and will be completed in later commits, the tests
should be disabled to avoid any temporary failures.

Update RecordService to remove the skipping of PM data and instead
remove the enricher calls from the PM event listener. This allows the
new PM POJO to be saved in the cache and database without triggering an
enrichment.

TIS21-4170
TIS21-2399