refactor: expand programme/curriculum to CM queue
When a Programme or Curriculum record is received the enrichment starts
from that record, performing a lookup of associated
CurriculumMemberships and syncing them all in a single job.

Refactor the enrichment to instead populate a queue with all related
CurriculumMemberships. This queue will then become the single source
from which records are pulled for CurriculumMembership enrichment.

This refactoring was previously done for the Placement enrichment.

TIS21-2399