feat: use db data for expanded PMs
When Programme Memberships are expanded from Programmes or Curriculum
Memberships the data that is queued is currently trusted as current.
However, it is possible for an update to already be in the queue so that
when the expanded PM gets read back from the queue the data in the
queued record and the database has diverged.

Mark expanded PMs with a new Operation value of `LOOKUP`, the programme
membership service should retrieve the latest data from the database
instead of overwritting it with stale data as happened when using the
`LOAD` operation.

TIS21-2399
TIS21-4172