fix: set message group ID for SNS publish
When pubishing to a FIFO SNS topic the Message Group ID must be set,
this defines which groups of messages must be procesed in order.
Create an ID for FIFO publishing which uses the schema, table and record
ID to ensure that all messages for the same data record are processed in
order. Taking this approach rather than having every message be in the
same group ensure that we can be efficient and not block updates that
can be handled in parallel.

TIS21-4085
TIS21-4076