feat: disable personal details update API calls
Personal details updates are now handled by messaging and the API
endpoints are to be removed. Disable to the calls to those API
endpoints.

`Person`/`basic-details` being emitted as an event has been rolled back
due to additional behaviour that needs to be account for, specifically
the deletion of the trainee profile.

TIS21-4085
TIS21-4076