feat: placement request/expansion for other sites
Placements should be expanded in to the queue, or requested when missing
based on PlacementSite relationships.

Create an event listener for placement sites which will expand
placements found using the related placementId, or alternatively request
the missing placement when not found. A deleted PlacementSite should
also trigger a re-sync of the Placement by re-queuing it to be synced.

Update the site event listener to also handle the placements related to
"other sites", using the PlacementSite records as a joining record.

TIS21-4969
TIS21-4396