feat: include block indemnity in PM enrichment
The specialty associated with a Programme Membership includes a
`blockIndemnity` flag. This flag should be included in the PM data when
it is enriched.

Update the AggregateCurriculumMembershDto and associated mapping
functions to handle the `blockIndemnity` field. It is a `tinybit` so the
raw data in the record will be `1` or `0` when present, and will need
converting to a boolean.

TIS21-5617
TIS21-5849