fix: carry training pathway to trainee details
The training pathway is lost due to not being mapped back to the
TraineeDetailsDto.
Update the DTO to add the training pathway and update the associated
mapper configs to ensure it is populated.

Add a new test to TcsSyncServiceTest to verify that programme
membership data is patched as expected.

TIS21-6175
TIS21-6181