fix: do not requeue old placement data
When placement related data is deleted the placement data is retrieved
from the database and queued to be processed.
If the queue already has a placement update or delete message then we
will process the new details, before then processing the old details.

Change the operation type from `LOAD` to `LOOKUP` so we always grab the
latest data available in the database. This will remove the ability for
overwriting data with old data, but we will still get duplicate effort
as we will process the placement twice. This should be considered
acceptable until the sync process can be overhauled.

TIS21-6701