fix(placement): always trigger downstream deletion
Currently if a `DELETE` message is received but no matching placement is
found in the sync database the `AfterDeleteEvent` listener does not fire
and downstream services are not notified of the potential deletion.

In most cases this is fine, as the deletion has already occured, but in
edge cases where the DELETE message is required again (e.g. cleaning up
partial/leftover data from previously failed deletes) to message never
makes it to the downstreams services leaving them out of date with no
way to automatically sync back up.

The original intention of the PlacementEventListener was that it would
fire on database events, allowing async handling of updates and deletes
de-coupled from the incoming record messages. However, the
implementation uses application events instead, which offer no such
benefits. As a result the event listeners offer nothing other than
complexity/obscurity given the current needs.

Delete the PlacementEventListener and update the PlacementSyncService to
call the placement enrichment directly. This simplifies the application
flow as well as removes the need for cache trickery to keep data
available after deletion.

TIS21-6701