feat: include DBC in programme membership
The admin endpoints for LTFT forms require filtering based on the
admin's local office, which can be determined using the `cognito:groups`
from the auth token.
The group name is the DBC, rather than a name, so to avoid having to
hardcode a mapping the code should be added to the PM so it can be
referenced in the LTFT data.

TIS21-6599
TIS21-6938